Move EXIT errors into from kvm.go into kvm/error.go
